web-development/aplication/requirements.txt
aniso8601==7.0.0
bcrypt==3.1.7
cffi==1.12.3
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
flask-marshmallow==0.10.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
marshmallow==3.0.1
pycparser==2.19
PyJWT==1.4.2
PyMySQL==0.9.3
pytz==2019.2
six==1.12.0
SQLAlchemy==1.3.8
Werkzeug==0.15.5
